+++ src/main/java/com/yanny/ytech/compatibility/jei/AlloyingRecipeCategory.java
@@ -3,21 +3,21 @@ package com.yanny.ytech.compatibility.jei;
 import com.yanny.ytech.YTechMod;
 import com.yanny.ytech.configuration.Utils;
 import com.yanny.ytech.configuration.recipe.AlloyingRecipe;
+import com.yanny.ytech.registration.YTechBlocks;
 import com.yanny.ytech.registration.YTechItems;
 import com.yanny.ytech.registration.YTechRecipeTypes;
 import mezz.jei.api.constants.RecipeTypes;
+import mezz.jei.api.constants.VanillaTypes;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
-import mezz.jei.api.gui.drawable.IDrawable;
+import mezz.jei.api.gui.builder.ITooltipBuilder;
 import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
+import mezz.jei.api.gui.widgets.IRecipeExtrasBuilder;
 import mezz.jei.api.helpers.IGuiHelper;
 import mezz.jei.api.recipe.IFocusGroup;
 import mezz.jei.api.recipe.RecipeIngredientRole;
 import mezz.jei.api.recipe.RecipeType;
-import mezz.jei.api.recipe.category.IRecipeCategory;
+import mezz.jei.api.recipe.category.AbstractRecipeCategory;
 import mezz.jei.api.registration.IRecipeCatalystRegistration;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.Font;
-import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
@@ -28,62 +28,50 @@ import org.jetbrains.annotations.NotNull;
 import java.util.Arrays;
 import java.util.List;
 
-public class AlloyingRecipeCategory implements IRecipeCategory<AlloyingRecipe> {
+public class AlloyingRecipeCategory extends AbstractRecipeCategory<AlloyingRecipe> {
     public static final RecipeType<AlloyingRecipe> RECIPE_TYPE = RecipeType.create(YTechMod.MOD_ID, "alloying", AlloyingRecipe.class);
-
-    private final Font font = Minecraft.getInstance().font;
-    private final IDrawable background;
-    private final IDrawable icon;
-    private final Component localizedName;
+    public static final ResourceLocation TEXTURE = Utils.modLoc("textures/gui/emi.png");
 
     public AlloyingRecipeCategory(IGuiHelper guiHelper) {
-        ResourceLocation location = Utils.modLoc("textures/gui/jei.png");
-        background = guiHelper.createDrawable(location, 82, 0, 92, 62);
-        icon = guiHelper.createDrawableItemStack(new ItemStack(YTechItems.PRIMITIVE_ALLOY_SMELTER.get()));
-        localizedName = Component.translatable("emi.category.ytech.alloying");
-    }
-
-    @NotNull
-    @Override
-    public RecipeType<AlloyingRecipe> getRecipeType() {
-        return RECIPE_TYPE;
-    }
-
-    @NotNull
-    @Override
-    public Component getTitle() {
-        return localizedName;
-    }
-
-    @NotNull
-    @Override
-    public IDrawable getBackground() {
-        return background;
-    }
-
-    @NotNull
-    @Override
-    public IDrawable getIcon() {
-        return icon;
+        super(
+                RECIPE_TYPE,
+                Component.translatable("emi.category.ytech.alloying"),
+                guiHelper.createDrawableItemLike(YTechItems.PRIMITIVE_ALLOY_SMELTER.get()),
+                112, 38
+        );
     }
 
     @Override
     public void setRecipe(IRecipeLayoutBuilder builder, AlloyingRecipe recipe, @NotNull IFocusGroup focuses) {
-        builder.addSlot(RecipeIngredientRole.INPUT, 1, 5).addItemStacks(Arrays.stream(recipe.ingredient1().getItems()).peek(i -> i.setCount(recipe.ingredient1().count())).toList());
-        builder.addSlot(RecipeIngredientRole.INPUT, 21, 5).addItemStacks(Arrays.stream(recipe.ingredient2().getItems()).peek(i -> i.setCount(recipe.ingredient2().count())).toList());
-        builder.addSlot(RecipeIngredientRole.OUTPUT, 71,  23).addItemStack(recipe.result());
+        builder.addSlot(RecipeIngredientRole.INPUT, 10, 5)
+                .setStandardSlotBackground()
+                .addIngredients(VanillaTypes.ITEM_STACK, Arrays.asList(recipe.ingredient1().getItems()));
+        builder.addSlot(RecipeIngredientRole.INPUT, 28, 5)
+                .setStandardSlotBackground()
+                .addIngredients(VanillaTypes.ITEM_STACK, Arrays.asList(recipe.ingredient2().getItems()));
+        builder.addSlot(RecipeIngredientRole.OUTPUT, 91,  5)
+                .setOutputSlotBackground()
+                .addItemStack(recipe.result());
     }
 
     @Override
-    public void draw(@NotNull AlloyingRecipe recipe, @NotNull IRecipeSlotsView recipeSlotsView, @NotNull GuiGraphics guiGraphics, double mouseX, double mouseY) {
-        String smeltingText = (recipe.smeltingTime() / 20) + "s";
-        String temperatureText = recipe.minTemperature() + "°C";
+    public void createRecipeExtras(@NotNull IRecipeExtrasBuilder builder, @NotNull AlloyingRecipe recipe, @NotNull IFocusGroup focuses) {
+        builder.addAnimatedRecipeArrow(recipe.smeltingTime()).setPosition(54, 5);
+        builder.addAnimatedRecipeFlame(1600).setPosition(19, 24);
+        builder.addWidget(new TemperatureWidget(0, 0, 4000));
+    }
 
-        int stringWidth = font.width(smeltingText);
-        guiGraphics.drawString(font, smeltingText, getWidth() - stringWidth, getHeight() - 8, 0xFF808080, false);
+    @Override
+    public void getTooltip(@NotNull ITooltipBuilder tooltip, @NotNull AlloyingRecipe recipe, @NotNull IRecipeSlotsView recipeSlotsView, double mouseX, double mouseY) {
+        super.getTooltip(tooltip, recipe, recipeSlotsView, mouseX, mouseY);
 
-        stringWidth = font.width(temperatureText);
-        guiGraphics.drawString(font, temperatureText, getWidth() - stringWidth, 0, 0xFF808080, false);
+        if (mouseX >= 54 && mouseX <= 54 + 24 && mouseY >= 5 && mouseY <= 5 + 17) {
+            tooltip.add(Component.translatable("emi.alloying.time", recipe.smeltingTime() / 20.0F));
+        }
+
+        if (mouseX >= 0 && mouseX <= 8 && mouseY >= 0 && mouseY <= 38) {
+            tooltip.add(Component.translatable("emi.alloying.temperature", recipe.minTemperature()));
+        }
     }
 
     public static List<AlloyingRecipe> getRecipes(@NotNull RecipeManager recipeManager) {
@@ -91,8 +79,8 @@ public class AlloyingRecipeCategory implements IRecipeCategory<AlloyingRecipe> {
     }
 
     public static void registerCatalyst(@NotNull IRecipeCatalystRegistration registration) {
-        registration.addRecipeCatalyst(new ItemStack(YTechItems.PRIMITIVE_ALLOY_SMELTER.get()), RECIPE_TYPE, RecipeTypes.FUELING);
-        registration.addRecipeCatalyst(new ItemStack(YTechItems.BRICK_CHIMNEY.get()), RECIPE_TYPE);
-        registration.addRecipeCatalyst(new ItemStack(YTechItems.REINFORCED_BRICK_CHIMNEY.get()), RECIPE_TYPE);
+        registration.addRecipeCatalyst(new ItemStack(YTechBlocks.PRIMITIVE_ALLOY_SMELTER.get()), RECIPE_TYPE, RecipeTypes.FUELING);
+        registration.addRecipeCatalyst(new ItemStack(YTechBlocks.BRICK_CHIMNEY.get()), RECIPE_TYPE);
+        registration.addRecipeCatalyst(new ItemStack(YTechBlocks.REINFORCED_BRICK_CHIMNEY.get()), RECIPE_TYPE);
     }
 }