+++ src/main/java/com/yanny/ytech/compatibility/jei/HammeringRecipeCategory.java
@@ -1,20 +1,19 @@
 package com.yanny.ytech.compatibility.jei;
 
 import com.yanny.ytech.YTechMod;
-import com.yanny.ytech.configuration.Utils;
 import com.yanny.ytech.configuration.recipe.HammeringRecipe;
+import com.yanny.ytech.registration.YTechBlocks;
 import com.yanny.ytech.registration.YTechItems;
 import com.yanny.ytech.registration.YTechRecipeTypes;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
-import mezz.jei.api.gui.drawable.IDrawable;
+import mezz.jei.api.gui.widgets.IRecipeExtrasBuilder;
 import mezz.jei.api.helpers.IGuiHelper;
 import mezz.jei.api.recipe.IFocusGroup;
 import mezz.jei.api.recipe.RecipeIngredientRole;
 import mezz.jei.api.recipe.RecipeType;
-import mezz.jei.api.recipe.category.IRecipeCategory;
+import mezz.jei.api.recipe.category.AbstractRecipeCategory;
 import mezz.jei.api.registration.IRecipeCatalystRegistration;
 import net.minecraft.network.chat.Component;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.RecipeHolder;
 import net.minecraft.world.item.crafting.RecipeManager;
@@ -22,49 +21,34 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
-public class HammeringRecipeCategory implements IRecipeCategory<HammeringRecipe> {
+public class HammeringRecipeCategory extends AbstractRecipeCategory<HammeringRecipe> {
     public static final RecipeType<HammeringRecipe> RECIPE_TYPE = RecipeType.create(YTechMod.MOD_ID, "hammering", HammeringRecipe.class);
 
-    private final IDrawable background;
-    private final IDrawable icon;
-    private final Component localizedName;
-
     public HammeringRecipeCategory(IGuiHelper guiHelper) {
-        ResourceLocation location = Utils.modLoc("textures/gui/jei.png");
-        background = guiHelper.createDrawable(location, 0, 86, 82, 52);
-        icon = guiHelper.createDrawableItemStack(new ItemStack(YTechItems.BRONZE_ANVIL.get()));
-        localizedName = Component.translatable("emi.category.ytech.hammering");
-    }
-
-    @NotNull
-    @Override
-    public RecipeType<HammeringRecipe> getRecipeType() {
-        return RECIPE_TYPE;
-    }
-
-    @NotNull
-    @Override
-    public Component getTitle() {
-        return localizedName;
-    }
-
-    @NotNull
-    @Override
-    public IDrawable getBackground() {
-        return background;
-    }
-
-    @NotNull
-    @Override
-    public IDrawable getIcon() {
-        return icon;
+        super(
+                RECIPE_TYPE,
+                Component.translatable("emi.category.ytech.hammering"),
+                guiHelper.createDrawableItemLike(YTechItems.BRONZE_ANVIL.get()),
+                84, 41
+        );
     }
 
     @Override
     public void setRecipe(IRecipeLayoutBuilder builder, HammeringRecipe recipe, @NotNull IFocusGroup focuses) {
-        builder.addSlot(RecipeIngredientRole.INPUT, 1, 9).addIngredients(recipe.ingredient());
-        builder.addSlot(RecipeIngredientRole.OUTPUT, 61,  9).addItemStack(recipe.result());
-        builder.addSlot(RecipeIngredientRole.CATALYST, 29,  31).addIngredients(recipe.tool());
+        builder.addSlot(RecipeIngredientRole.INPUT, 0, 5)
+                .setStandardSlotBackground()
+                .addIngredients(recipe.ingredient());
+        builder.addSlot(RecipeIngredientRole.CATALYST, 29,  24)
+                .setStandardSlotBackground()
+                .addIngredients(recipe.tool());
+        builder.addSlot(RecipeIngredientRole.OUTPUT, 62,  5)
+                .setOutputSlotBackground()
+                .addItemStack(recipe.result());
+    }
+
+    @Override
+    public void createRecipeExtras(IRecipeExtrasBuilder builder, @NotNull HammeringRecipe recipe, @NotNull IFocusGroup focuses) {
+        builder.addRecipeArrow().setPosition(26, 5);
     }
 
     public static List<HammeringRecipe> getRecipes(@NotNull RecipeManager recipeManager) {
@@ -72,6 +56,6 @@ public class HammeringRecipeCategory implements IRecipeCategory<HammeringRecipe>
     }
 
     public static void registerCatalyst(@NotNull IRecipeCatalystRegistration registration) {
-        registration.addRecipeCatalyst(new ItemStack(YTechItems.BRONZE_ANVIL.get()), RECIPE_TYPE);
+        registration.addRecipeCatalyst(new ItemStack(YTechBlocks.BRONZE_ANVIL.get()), RECIPE_TYPE);
     }
 }