+++ src/main/java/com/yanny/ytech/compatibility/jei/TanningRecipeCategory.java
@@ -2,24 +2,20 @@ package com.yanny.ytech.compatibility.jei;
 
 import com.yanny.ytech.YTechMod;
 import com.yanny.ytech.configuration.MaterialType;
-import com.yanny.ytech.configuration.Utils;
 import com.yanny.ytech.configuration.recipe.TanningRecipe;
 import com.yanny.ytech.registration.YTechItems;
 import com.yanny.ytech.registration.YTechRecipeTypes;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
-import mezz.jei.api.gui.drawable.IDrawable;
+import mezz.jei.api.gui.builder.ITooltipBuilder;
 import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
+import mezz.jei.api.gui.widgets.IRecipeExtrasBuilder;
 import mezz.jei.api.helpers.IGuiHelper;
 import mezz.jei.api.recipe.IFocusGroup;
 import mezz.jei.api.recipe.RecipeIngredientRole;
 import mezz.jei.api.recipe.RecipeType;
-import mezz.jei.api.recipe.category.IRecipeCategory;
+import mezz.jei.api.recipe.category.AbstractRecipeCategory;
 import mezz.jei.api.registration.IRecipeCatalystRegistration;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.Font;
-import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.network.chat.Component;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.RecipeHolder;
 import net.minecraft.world.item.crafting.RecipeManager;
@@ -27,55 +23,41 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
-public class TanningRecipeCategory implements IRecipeCategory<TanningRecipe> {
+public class TanningRecipeCategory extends AbstractRecipeCategory<TanningRecipe> {
     public static final RecipeType<TanningRecipe> RECIPE_TYPE = RecipeType.create(YTechMod.MOD_ID, "tanning", TanningRecipe.class);
 
-    private final Font font = Minecraft.getInstance().font;
-    private final IDrawable background;
-    private final IDrawable icon;
-    private final Component localizedName;
-
     public TanningRecipeCategory(IGuiHelper guiHelper) {
-        ResourceLocation location = Utils.modLoc("textures/gui/jei.png");
-        background = guiHelper.createDrawable(location, 0, 34, 82, 52);
-        icon = guiHelper.createDrawableItemStack(new ItemStack(YTechItems.TANNING_RACKS.get(MaterialType.OAK_WOOD).get()));
-        localizedName = Component.translatable("emi.category.ytech.tanning");
-    }
-
-    @NotNull
-    @Override
-    public RecipeType<TanningRecipe> getRecipeType() {
-        return RECIPE_TYPE;
-    }
-
-    @NotNull
-    @Override
-    public Component getTitle() {
-        return localizedName;
-    }
-
-    @NotNull
-    @Override
-    public IDrawable getBackground() {
-        return background;
-    }
-
-    @NotNull
-    @Override
-    public IDrawable getIcon() {
-        return icon;
+        super(
+                RECIPE_TYPE,
+                Component.translatable("emi.category.ytech.tanning"),
+                guiHelper.createDrawableItemLike(YTechItems.TANNING_RACKS.get(MaterialType.OAK_WOOD).get()),
+                84, 41
+        );
     }
 
     @Override
     public void setRecipe(IRecipeLayoutBuilder builder, TanningRecipe recipe, @NotNull IFocusGroup focuses) {
-        builder.addSlot(RecipeIngredientRole.INPUT, 1, 9).addIngredients(recipe.ingredient());
-        builder.addSlot(RecipeIngredientRole.OUTPUT, 61,  9).addItemStack(recipe.result());
-        builder.addSlot(RecipeIngredientRole.CATALYST, 29,  31).addIngredients(recipe.tool());
+        builder.addSlot(RecipeIngredientRole.INPUT, 0, 5)
+                .setStandardSlotBackground()
+                .addIngredients(recipe.ingredient());
+        builder.addSlot(RecipeIngredientRole.CATALYST, 29,  24)
+                .setStandardSlotBackground()
+                .addIngredients(recipe.tool());
+        builder.addSlot(RecipeIngredientRole.OUTPUT, 62,  5)
+                .setOutputSlotBackground()
+                .addItemStack(recipe.result());
     }
 
     @Override
-    public void draw(@NotNull TanningRecipe recipe, @NotNull IRecipeSlotsView recipeSlotsView, @NotNull GuiGraphics guiGraphics, double mouseX, double mouseY) {
-        guiGraphics.drawString(font, recipe.hitCount() + "x", 48, 35, 0xFF808080, false);
+    public void createRecipeExtras(IRecipeExtrasBuilder builder, @NotNull TanningRecipe recipe, @NotNull IFocusGroup focuses) {
+        builder.addRecipeArrow().setPosition(26, 5);
+    }
+
+    @Override
+    public void getTooltip(@NotNull ITooltipBuilder tooltip, @NotNull TanningRecipe recipe, @NotNull IRecipeSlotsView recipeSlotsView, double mouseX, double mouseY) {
+        if (mouseX >= 26 && mouseX <= 26 + 24 && mouseY >= 5 && mouseY <= 5 + 17) {
+            tooltip.add(Component.translatable("emi.tanning.hit_count", recipe.hitCount()));
+        }
     }
 
     public static List<TanningRecipe> getRecipes(@NotNull RecipeManager recipeManager) {