+++ src/main/java/com/yanny/ytech/registration/YTechBlocks.java
@@ -42,6 +42,7 @@ public class YTechBlocks {
     public static final DeferredBlock<Block> THATCH_SLAB = registerSlab("thatch_slab", Blocks.HAY_BLOCK);
     public static final DeferredBlock<Block> THATCH_STAIRS = registerStairs("thatch_stairs", YTechBlocks.THATCH, Blocks.HAY_BLOCK);
     public static final DeferredBlock<Block> TREE_STUMP = BLOCKS.register("tree_stump", TreeStumpBlock::new);
+    public static final DeferredBlock<Block> WOODEN_BOX = BLOCKS.register("wooden_box", WoodenBoxBlock::new);
 
     public static final MaterialBlock DEEPSLATE_ORES = new DeepslateOreMaterialBlock();
     public static final MaterialBlock DRYING_RACKS = new MaterialBlock("drying_rack", NameHolder.suffix("drying_rack"), MaterialType.ALL_WOODS, DryingRackBlock::new);